pyproject.toml
[project]
name = "openapi-test-client"
version = "0.1"
requires-python = ">=3.11"
dependencies = [
    "autoflake==2.3.0",
    "black==23.12.1",
    "isort==5.12.0",
    "inflect==7.0.0",
    "pre-commit==3.6.2",
    "pydantic[email]==2.6.2",
    "PyYAML==6.0.1",
    "requests==2.31.0"
]
readme = "README.md"

[project.optional-dependencies]
app = [
    "Quart==0.19.4",
    "quart-auth==0.9.0",
    "quart-schema[pydantic]==0.19.1",

]
test = [
    "openapi-test-client[app]",
    "pytest==7.4.4",
    "pytest-subtests==0.11.0",
]

[build-system]
requires = ["setuptools >= 61.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
openapi-client = "openapi_test_client.scripts.generate_client:main"

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true
profile = "black"

[tool.black]
line_length = 120
